requirements.txt
Mako==1.3.10
MarkupSafe==3.0.3
PyJWT==2.10.1
Pygments==2.19.2
SQLAlchemy==2.0.45
alembic==1.18.4
alembic-postgresql-enum==0.1.9
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
boto3==1.42.12
botocore==1.42.34
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.44.1
fastapi==0.129.0
fastapi-storage==1.0.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jmespath==1.1.0
loguru==0.7.3
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-instrumentation-sqlalchemy==0.60b1
opentelemetry-sdk==1.39.1
packaging==26.0
pgvector==0.4.2
pluggy==1.6.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-extra-types==2.10.6
pydantic_core==2.41.5
pytest==9.0.2
pytest-dotenv==0.5.2
pytest-fastapi==0.1.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python-logging-loki==0.3.1
python-multipart==0.0.22
pytz==2025.2
requests==2.32.5
ruff==0.14.9
s3transfer==0.16.0
six==1.17.0
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.38.0